+++ framework/py/flwr/compat/client/app.py
@@ -37,13 +37,13 @@ from flwr.common import GRPC_MAX_MESSAGE_LENGTH, EventType, event
 from flwr.common.constant import MAX_RETRY_DELAY, ErrorCode
 from flwr.common.exit import ExitCode, flwr_exit
 from flwr.common.logger import log, warn_deprecated_feature
-from flwr.common.retry_invoker import RetryInvoker, RetryState, exponential
 from flwr.compat.client.grpc_client.connection import grpc_connection
 from flwr.compat.client.typing import ClientFnExt
 from flwr.compat.common.constant import TRANSPORT_TYPE_GRPC_BIDI, TRANSPORT_TYPES_COMPAT
 from flwr.supercore.address import parse_address
 from flwr.supercore.fab import Fab
 from flwr.supercore.object_store import ObjectStoreFactory
+from flwr.supercore.retry import RetryInvoker, RetryState, exponential
 from flwr.supercore.run import Run, RunNotRunningException
 from flwr.supernode.nodestate import NodeStateFactory
