+++ framework/py/flwr/supernode/start_client_internal.py
@@ -52,7 +52,6 @@ from flwr.common.constant import (
 from flwr.common.exit import ExitCode, flwr_exit, register_signal_handlers
 from flwr.common.grpc import generic_create_grpc_server
 from flwr.common.logger import log
-from flwr.common.retry_invoker import RetryInvoker, make_simple_grpc_retry_invoker
 from flwr.common.telemetry import EventType
 from flwr.proto.clientappio_pb2_grpc import add_ClientAppIoServicer_to_server
 from flwr.proto.message_pb2 import ObjectTree  # pylint: disable=E0611
@@ -81,6 +80,7 @@ from flwr.supercore.primitives.asymmetric_ed25519 import (
     decode_base64url,
     verify_signature,
 )
+from flwr.supercore.retry import RetryInvoker, make_simple_grpc_retry_invoker
 from flwr.supercore.run import Run, RunNotRunningException
 from flwr.supercore.tls import get_client_tls_args
 from flwr.supercore.version import package_version
